refactor(study-search): move filter helpers to separate module
